+++ trunk/bpel-g/org.activebpel.rt.bpel/src/main/java/org/activebpel/rt/bpel/def/util/AeLocationPathUtils.java
@@ -52,7 +52,7 @@ public class AeLocationPathUtils
             try
             {
                Integer instInt = Integer.valueOf(instStr);
-               return instInt.intValue();
+               return instInt;
             }
             catch (NumberFormatException e)
             {
