+++ trunk/bpel-g/org.activebpel.rt.bpeladmin.jar/src/main/java/org/activebpel/rt/bpeladmin/war/web/AeProcessInstanceDetailWrapper.java
@@ -160,7 +160,7 @@ public class AeProcessInstanceDetailWrapper
       long pid = -1;
       try
       {
-         pid = Long.valueOf(aPid).longValue();
+         pid = Long.valueOf(aPid);
       }
       catch (Exception ex)
       {
